requirements.txt
confluent-kafka>=2.0.2
datasets>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.30.0
shap>=0.42.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.13.0
requests>=2.31.0
prometheus-client>=0.18.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
